CodebreakerApplication (example 5 - red) Feature: Code-breaker starts game Scenario: Start game Step: I should see "Welcome to Codebreaker!" - red Specification: CodebreakerApplication Example: it_should_avoid_expecting_a_command_name - red
To finish our single command application we need to avoid expecting a command name, the way to do that is clear out the normal first argument and to accomplish that we need to be sure the method `setArguments` of the `InputDefinition` class is called inside the `getDefinition` method. So this is calling for a Mock! the method to achieve this is the next one

    use Symfony\Component\Console\Input\InputDefinition;

    function it_should_avoid_expecting_a_command_name(InputDefinition $definition)
    {
        $definition->setArguments()->shouldBeCalledTimes(1);

        $this->setDefinition($definition);
        $this->getDefinition();
    }

Again we have a failing example